requirements.txt
numpy
pandas
scikit-learn
tensorflow>=2.10
joblib
Pillow
matplotlib
streamlit
kaggle
tqdm